move existing rpc tests into mocked module
