+++ rpc/src/v1/tests/mod.rs
@@ -1,31 +1,8 @@
-// Copyright 2015, 2016 Ethcore (UK) Ltd.
-// This file is part of Parity.
-
-// Parity is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Parity is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
-
-//!TODO: load custom blockchain state and test
+//! RPC unit test moduleS
 
 pub mod helpers;
+
+#[cfg(test)]
+mod mocked;
 #[cfg(test)]
 mod eth;
-#[cfg(test)]
-mod net;
-#[cfg(test)]
-mod web3;
-#[cfg(test)]
-mod personal;
-#[cfg(test)]
-mod ethcore;
-#[cfg(test)]
-mod rpc;
